Version 7.1.39 - Added extreme optimization improvements for all player/combat-related tasks - Big rewrite of numerous aspects of the combat system, but no changes in how it should behave (optimizations only) - Players are now able to pick up Elite Coins even when their inventories are full - Elite Coins now follow the player until they get picked up - Bosses now drop more Elite Coins if they are higher level than the player's gear and fewer coins if they are lower level, up to a point where they drop no coins (doesn't affect other loot) - Fixed prestige'd custom rank names not appearing at places - Hunter enchantments now only work in armor slots - Critical enchantments now only work for weapons - Removed VanishNoPacket compatibility - Increased party bonus for natural mob spawning - Natural spawning system no longer scans entire hotbar, only held item for natural mob spawn - Simplified the natural mob spawning logic, tweaked some custom boss dynamic level logic - Fixed duping bug related to trying to wear soulbound armor that belongs to someone else - Fixed NPE related to tracking bosses that no longer exist - Tentative fix for issue related to setting database values - Fixed a bug where regional bosses would never respawn for users of some java version

+++ src/main/java/com/magmaguy/elitemobs/combatsystem/CombatSystem.java
@@ -8,12 +8,12 @@ public class CombatSystem {
     public static final double PER_LEVEL_POWER_INCREASE = 0.5;
     public static final double TARGET_HITS_TO_KILL = 7; //affects max health assignment on EliteMobEntity.java
 
-    public static final double TRIDENT = 9;
-    public static final double NETHERITE_TIER_LEVEL = 8;
-    public static final double DIAMOND_TIER_LEVEL = 7;
-    public static final double IRON_TIER_LEVEL = 6;
-    public static final double STONE_CHAIN_TIER_LEVEL = 5;
-    public static final double GOLD_WOOD_LEATHER_TIER_LEVEL = 3;
+    public static final int TRIDENT = 9;
+    public static final int NETHERITE_TIER_LEVEL = 8;
+    public static final int DIAMOND_TIER_LEVEL = 7;
+    public static final int IRON_TIER_LEVEL = 6;
+    public static final int STONE_CHAIN_TIER_LEVEL = 5;
+    public static final int GOLD_WOOD_LEATHER_TIER_LEVEL = 3;
 
     public static boolean bypass = false;
